refactor(onsite_projector): 拆分打印函数到独立文件，使用下划线命名法
